Improved @ban command
When banning a character, the game master now sets a time unit (m, h, d, w
or y for minutes, hours, days, weeks or years) after the duration.
Ban durations longer than 2^16 minutes are now possible.

The banned character is now kicked automatically and the banning character
receives a feedback chat message.

+++ src/account-server/serverhandler.cpp
@@ -383,7 +383,7 @@ void ServerHandler::processMessage(NetComputer *comp, MessageIn &msg)
         case GAMSG_BAN_PLAYER:
         {
             int id = msg.readInt32();
-            int duration = msg.readInt16();
+            int duration = msg.readInt32();
             storage->banCharacter(id, duration);
         } break;
